created AbstractAuthSessionManager as base class for OpenShiftAuthSessionManager, RedHatAuthSessionManager
