+++ src/main/kotlin/com/redhat/devtools/gateway/auth/session/AuthSessionManager.kt
@@ -23,6 +23,9 @@ interface AuthSessionManager {
     /** Starts login and returns browser URL */
     suspend fun startLogin(apiServerUrl: String? = null, sslContext: SSLContext): URI
 
+    /** Awaits for the browser login result */
+    suspend fun awaitLoginResult(timeoutMs: Long): SSOToken
+
     /** Starts login using the given credentials and returns a valid token */
     suspend fun loginWithCredentials(apiServerUrl: String, username: String, password: String, sslContext: SSLContext): SSOToken
